pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xauusd-bot"
version = "0.1.0"
description = "XAUUSD multi-timeframe simulator/backtest scaffold"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.1",
  "pyyaml>=6.0",
  "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
